pyproject.toml
[project]
name = "multi-system-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "asyncpg>=0.29.0",
    "httpx>=0.25.0",
    "mcp>=1.26.0",
    "groq>=0.4.2",
    "langchain-groq>=0.1.0",
    "langchain>=0.1.0",
    "langgraph>=0.2.0",
    "python-dotenv>=1.0.0"
]
